Remove ref/lock on clock's time field
And pass clock reference to modules (oscillator), instead of time.

The clock itself is already reference counted for synchronisation. It's
easier to just pass a reference of the entire clock to the modules that
need it, instead of further having to lock one of it's values.

Also, chnange SAMPLE_RATE to be f32 as it's only used with f32 values

+++ yat-rack/src/types.rs
@@ -3,7 +3,7 @@ use std::fmt;
 use std::error::Error;
 
 pub type SampleType = f32;
-pub const SAMPLE_RATE: u32 = 44100;
+pub const SAMPLE_RATE: f32 = 44100f32;
 pub const AUDIO_BUF_SIZE: usize = 1024;
 pub type IoPort = Arc<RwLock<Option<SampleType>>>;
 pub type ModuleResult<T> = std::result::Result<T, ModuleNotFoundError>;
